+++ src/hotspot/share/runtime/globals_shared.hpp
@@ -37,7 +37,7 @@
 // parts of the memory system may require additional alignment
 // and are responsible for those alignments.
 #ifdef _LP64
-#define ScaleForWordSize(x) align_down((x) * 13 / 10, HeapWordSize)
+#define ScaleForWordSize(x) align_down((x) * 13 / 10, static_cast<size_t>(HeapWordSize))
 #else
 #define ScaleForWordSize(x) (x)
 #endif
